requirements.txt
aiocron==1.3
aiohttp==3.7.3
aiohttp-middlewares==1.1.0
alembic==1.5.2
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
chardet==3.0.4
colorama==0.4.4
coloredlogs==15.0
colorlog==4.7.2
croniter==1.0.2
gino==1.0.1
gino-aiohttp==0.2.0
gunicorn==20.0.4
humanfriendly==9.1
idna==3.1
iniconfig==1.1.1
loguru==0.5.3
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
natsort==7.1.1
packaging==20.8
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-aiohttp==0.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
uvloop==0.14.0
yarl==1.6.3
